Drop some mis-leading text.
Here, we sometimes mislead the reader.  I saw a message today that read:

> Event state is switched to COMPLETE due to "17 container image(s) failed to rebuild."

Why is it complete if 17 failed?  When I looked more closely, I found that
there were 40 images.  23 of them succeeded and 17 failed.  The
`state_reason` is really more like a `summary`, so let's expose it that
way to users.  # Please enter the commit message for your changes. Lines
starting