errata: add a subtitle catch-all clause, and additional tests
If a message had a value other than "null", "0", or "1" in the "to"
header, no subtitle template would be set, and an error would be thrown
when calling format(). This change adds a catch-all clause that generates
a generic subtitle for values it doesn't expect.

Tests for the catch-all case, and for security approval are also added.