set safe.directory so git will read files owned by a different user
The user running the git command is different than the user that owns the git checkout
because the Makefile is run inside a mock environment, with the git checkout bind-mounted
into it. "git log" is used to include the commit hash in the srpm Release value, but Git
refuses to read any files owned by a different user unless safe.checkout is set.